+++ src/main/java/com/halggeol/backend/survey/dto/KnowledgeSurveyRequestDTO.java
@@ -1,10 +1,8 @@
-package com.halggeol.backend.user.dto;
+package com.halggeol.backend.survey.dto;
 
 import java.util.List;
 import javax.validation.Valid;
 import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 import lombok.AllArgsConstructor;
