+++ src/main/java/com/halggeol/backend/user/controller/UserController.java
@@ -3,8 +3,6 @@ package com.halggeol.backend.user.controller;
 import com.halggeol.backend.security.domain.CustomUser;
 import com.halggeol.backend.user.dto.EditProfileDTO;
 import com.halggeol.backend.user.dto.EmailDTO;
-import com.halggeol.backend.user.dto.KnowledgeSurveyRequestDTO;
-import com.halggeol.backend.user.dto.TendencySurveyRequestDTO;
 import com.halggeol.backend.user.dto.UserJoinDTO;
 import com.halggeol.backend.user.service.UserService;
 import java.util.Map;
@@ -13,7 +11,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.lang.Nullable;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -32,20 +29,20 @@ import org.springframework.web.bind.annotation.RestController;
 public class UserController {
     private final UserService userService;
 
-    // 회원가입 요청 (이메일 본인 인증) API
     @PostMapping("/signup/request")
     public ResponseEntity<Map<String, String>> requestJoin(
         @Valid @RequestBody EmailDTO email
     ) {
+        // 회원가입 요청 (이메일 본인 인증)
         return ResponseEntity.ok(userService.requestJoin(email));
     }
 
-    // 회원가입 등록 API
     @PostMapping("/signup")
     public ResponseEntity<Map<String, String>> join(
         @Valid @RequestBody UserJoinDTO user,
         @RequestParam String token
     ) {
+        // 회원가입 등록 (회원 정보 저장)
         return ResponseEntity.status(HttpStatus.CREATED).body(userService.join(user, token));
     }
 
@@ -54,6 +51,7 @@ public class UserController {
         @AuthenticationPrincipal CustomUser user,
         @RequestParam(required = false) String scope
     ) {
+        // 마이페이지 회원 정보 조회
         return ResponseEntity.ok(userService.viewProfile(user, scope));
     }
 
@@ -62,6 +60,7 @@ public class UserController {
         @AuthenticationPrincipal CustomUser user,
         @Valid @RequestBody EditProfileDTO info
     ) {
+        // 마이페이지 회원 정보 수정
         return ResponseEntity.ok(userService.editProfile(user, info));
     }
 
@@ -70,20 +69,7 @@ public class UserController {
         @AuthenticationPrincipal CustomUser user,
         @RequestHeader("Authorization") String bearerToken
     ) {
+        // 회원 탈퇴
         return ResponseEntity.ok(userService.deleteUser(user, bearerToken));
-    }
-
-    @PatchMapping("/survey/knowledge")
-    public ResponseEntity<Map<String, String>> updateKnowledge(
-        @Valid @RequestBody KnowledgeSurveyRequestDTO surveyResult
-    ) {
-        return ResponseEntity.ok(userService.updateKnowledge(surveyResult));
-    }
-
-    @PatchMapping("/survey/tendency")
-    public ResponseEntity<Map<String, String>> updateTendency(
-        @Valid @RequestBody TendencySurveyRequestDTO surveyResult
-    ) {
-        return ResponseEntity.ok(userService.updateTendency(surveyResult));
     }
 }
