+++ src/main/java/com/halggeol/backend/user/mapper/UserMapper.java
@@ -33,20 +33,5 @@ public interface UserMapper {
         @Param("phone") String phone
     );
 
-    int updateKnowledgeByEmail(
-        @Param("email") String email,
-        @Param("userKlg") int userKlg
-    );
-
-    int updateTendencyByEmail(
-        @Param("email") String email,
-        @Param("risk") int risk,
-        @Param("yieldScore") double yieldScore,
-        @Param("riskScore") double riskScore,
-        @Param("costScore") double costScore,
-        @Param("liquidityScore") double liquidityScore,
-        @Param("complexityScore") double complexityScore
-    );
-
     int deleteUserById(int id);
 }
