+++ src/main/java/com/halggeol/backend/user/service/UserService.java
@@ -3,13 +3,10 @@ package com.halggeol.backend.user.service;
 import com.halggeol.backend.security.domain.CustomUser;
 import com.halggeol.backend.user.dto.EditProfileDTO;
 import com.halggeol.backend.user.dto.EmailDTO;
-import com.halggeol.backend.user.dto.KnowledgeSurveyRequestDTO;
-import com.halggeol.backend.user.dto.TendencySurveyRequestDTO;
+import com.halggeol.backend.survey.dto.KnowledgeSurveyRequestDTO;
+import com.halggeol.backend.survey.dto.TendencySurveyRequestDTO;
 import com.halggeol.backend.user.dto.UserJoinDTO;
 import java.util.Map;
-import javax.validation.Valid;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 
 public interface UserService {
     boolean findByEmail(String email);
@@ -25,8 +22,4 @@ public interface UserService {
     Map<String, String> editProfile(CustomUser user, EditProfileDTO info);
 
     Map<String, String> deleteUser(CustomUser user, String bearerToken);
-
-    Map<String, String> updateKnowledge(KnowledgeSurveyRequestDTO surveyResult);
-
-    Map<String, String> updateTendency(TendencySurveyRequestDTO surveyResult);
 }
