+++ src/main/java/com/halggeol/backend/user/service/UserServiceImpl.java
@@ -10,8 +10,8 @@ import com.halggeol.backend.security.util.JwtManager;
 import com.halggeol.backend.security.util.SecurityUtil;
 import com.halggeol.backend.user.dto.EditProfileDTO;
 import com.halggeol.backend.user.dto.EmailDTO;
-import com.halggeol.backend.user.dto.KnowledgeSurveyRequestDTO;
-import com.halggeol.backend.user.dto.TendencySurveyRequestDTO;
+import com.halggeol.backend.survey.dto.KnowledgeSurveyRequestDTO;
+import com.halggeol.backend.survey.dto.TendencySurveyRequestDTO;
 import com.halggeol.backend.user.dto.UserJoinDTO;
 import com.halggeol.backend.user.dto.UserProfileResponseDTO;
 import com.halggeol.backend.user.mapper.UserMapper;
@@ -108,46 +108,5 @@ public class UserServiceImpl implements UserService {
         }
         userMapper.deleteUserById(user.getUser().getId());
         return Map.of("Message", "회원탈퇴가 완료되었습니다.");
-    }
-
-    @Override
-    public Map<String, String> updateKnowledge(KnowledgeSurveyRequestDTO surveyResult) {
-        String email = securityUtil.resolveEmail(surveyResult.getEmail());
-        if (email == null) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "이메일이 입력되지 않았습니다.");
-        }
-        if (!findByEmail(email)) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "존재하지 않는 사용자입니다.");
-        }
-
-        int userKlg = 0; // surveyResult로 점수 내기
-
-        userMapper.updateKnowledgeByEmail(email, userKlg);
-        return Map.of("Message", "금융 이해도 갱신이 완료되었습니다.");
-    }
-
-    @Override
-    public Map<String, String> updateTendency(TendencySurveyRequestDTO surveyResult) {
-        String email = securityUtil.resolveEmail(surveyResult.getEmail());
-        if (email == null) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "이메일이 입력되지 않았습니다.");
-        }
-        if (!findByEmail(email)) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "존재하지 않는 사용자입니다.");
-        }
-
-        int risk = 0; // surveyResult로 점수 내기
-        UserVectorResponseDTO scores = null; // 5개 user score 점수 내기
-
-        userMapper.updateTendencyByEmail(
-            email,
-            risk,
-            scores.getYieldScore(),
-            scores.getRiskScore(),
-            scores.getCostScore(),
-            scores.getLiquidityScore(),
-            scores.getComplexityScore()
-        );
-        return Map.of("Message", "투자 성향 갱신이 완료되었습니다.");
     }
 }
